rename tenantcappuploader admin service module
