+++ modules/service-stubs/org.wso2.intcloud.services.tenant.carbonapps.stub/pom.xml
@@ -11,10 +11,10 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>org.wso2.intcloud.services.tenantcappuploader.stub</artifactId>
+    <artifactId>org.wso2.intcloud.services.tenant.carbonapps.stub</artifactId>
     <packaging>bundle</packaging>
     <name>WSO2 Integration Cloud Components - Tenant CApp Uploader Service Stub</name>
-    <description>org.wso2.intcloud.services.tenantcappuploader.stub</description>
+    <description>org.wso2.intcloud.services.tenant.carbonapps.stub</description>
 
     <url>http://wso2.org</url>
     <build>
@@ -27,9 +27,9 @@
                     <instructions>
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
-                        <Export-Package>org.wso2.intcloud.services.tenantcappuploader.*;
+                        <Export-Package>org.wso2.intcloud.services.tenant.carbonapps.*;
                             version="${appcloud.version}"</Export-Package>
-                        <Import-Package>!org.wso2.intcloud.services.tenantcappuploader.stub.*</Import-Package>
+                        <Import-Package>!org.wso2.intcloud.services.tenant.carbonapps.stub.*</Import-Package>
                         <DynamicImport-Package>*</DynamicImport-Package>
                     </instructions>
                 </configuration>
@@ -54,7 +54,7 @@
                                     <pathelement location="${settings.localRepository}/org/apache/ws/commons/axiom/wso2/axiom/${axiom.wso2.version}/axiom-${axiom.wso2.version}.jar" />
                                 </path>
                                 <java classname="org.apache.axis2.wsdl.WSDL2Java" fork="true">
-                                    <arg line="-uri src/main/resources/TenantCarbonAppUploader.wsdl -p org.wso2.intcloud.services.tenantcappuploader.stub.types -ns2p http://org.apache.axis2/xsd=org.wso2.intcloud.services.tenantcappuploader.stub.types.axis2,http://tenantcappuploader.services.intcloud.wso2.org/xsd=org.wso2.intcloud.services.tenantcappuploader.stub.types,http://io.java/xsd=java.io,http://exceptions.core.registry.carbon.wso2.org/xsd=org.wso2.carbon.registry.core.exceptions -u -uw -o target/generated-code" />
+                                    <arg line="-uri src/main/resources/TenantCarbonAppAdminService.wsdl -p org.wso2.intcloud.services.tenant.carbonapps.stub.types -ns2p http://org.apache.axis2/xsd=org.wso2.intcloud.services.tenant.carbonapps.stub.types.axis2,http://carbonapps.tenant.services.intcloud.wso2.org/xsd=org.wso2.intcloud.services.tenant.carbonapps.stub.types,http://io.java/xsd=java.io,http://exceptions.core.registry.carbon.wso2.org/xsd=org.wso2.carbon.registry.core.exceptions -u -uw -o target/generated-code" />
                                     <classpath refid="wsdl2java.classpath" />
                                     <classpath refid="maven.dependency.classpath" />
                                     <classpath refid="maven.compile.classpath" />
