+++ modules/services/org.wso2.intcloud.services.tenant.carbonapps/pom.xml
@@ -11,10 +11,10 @@
         <relativePath>../../../pom.xml</relativePath>
     </parent>
 
-    <artifactId>org.wso2.intcloud.services.tenantcappuploader</artifactId>
+    <artifactId>org.wso2.intcloud.services.tenant.carbonapps</artifactId>
     <packaging>bundle</packaging>
-    <name>WSO2 Integration Cloud Services - Tenant Carbon Application Uploader Service</name>
-    <description>org.wso2.intcloud.services.tenantcappuploader</description>
+    <name>WSO2 Integration Cloud Services - Tenant Carbon Application Service</name>
+    <description>org.wso2.intcloud.services.tenant.carbonapps</description>
 
     <dependencies>
         <dependency>
@@ -157,8 +157,8 @@
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Export-Package>
-                            org.wso2.intcloud.services.tenantcappuploader.*,
-                            !org.wso2.intcloud.services.tenantcappuploader
+                            org.wso2.intcloud.services.tenant.carbonapps.deployer.*,
+                            !org.wso2.intcloud.services.tenant.carbonapps.deployer
                         </Export-Package>
                         <Import-Package>
                             org.wso2.carbon.registry.core.*;version=1.0.1,
