+++ modules/services/org.wso2.intcloud.services.tenant.carbonapps/src/main/resources/META-INF/services.xml
@@ -16,12 +16,12 @@
  ! limitations under the License.
  !-->
 <serviceGroup>
-    <service name="TenantCarbonAppUploader" scope="transportsession">
+    <service name="TenantCarbonAppAdminService" scope="transportsession">
         <transports>
             <transport>https</transport>
         </transports>
         <parameter name="ServiceClass" locked="false">
-            org.wso2.intcloud.services.tenantcappuploader.TenantCarbonAppUploader
+            org.wso2.intcloud.services.tenant.carbonapps.deployer.TenantCarbonAppUploader
         </parameter>
     </service>
     <parameter name="AuthorizationAction" locked="true">/permission/admin/login</parameter>
